mslearn-ai-document-intelligence/Labfiles/05-content-understanding/code/requirements.txt
python-dotenv>=1.0.0
azure-identity
azure-ai-projects==2.0.0b4
requests>=2.31.0
